requirements.txt
google-cloud-texttospeech
requests
beautifulsoup4
pygame==2.5.2